fixed circular import possibility & dnpdata config
added config for max_print_attrs and print_values
