+++ dnplab/config/dnplab_cfg.cfg
@@ -7,6 +7,11 @@ light_green = #67AE3E
 light_grey = #A7A9AC
 #black = #000000
 
+[CORE]
+
+DNPData_max_print_attrs = 5
+DNPData_print_values = False
+
 [PLOTTING]
 
 forwarded_pyplot_plots = [ semilogy, semilogx, polar, loglog, scatter, errorbar, step ]
